segmentation/src/requirements.txt
torch
torchvision
albumentations
opencv-python
pandas
numpy
pyyaml
tqdm